updated database context
